+++ test/vonage/test.rb
@@ -114,7 +114,9 @@ module Vonage
 
     def request(body: nil, query: nil, headers: {}, auth: nil)
       headers['Authorization'] = auth || authorization
-      headers['Content-Type'] = 'application/json' if body
+      if body
+        headers['Content-Type'] = 'application/json' unless headers.has_key?('Content-Type')
+      end
 
       { headers: headers, body: body, query: query }.compact
     end
@@ -354,6 +356,22 @@ module Vonage
         body:
           '{"_embedded":[{"key":"value"}, {"key":"value"}, {"key":"value"}]}'
       }
+    end
+
+    def network_authentication_auth_request_id
+      'b1963d15-537f-459a-be89-e00fc310b82b'
+    end
+
+    def network_authentication_token_response
+      { 
+        body: '{"access_token":"' + sample_webhook_token + '", "token_type":"Bearer", "expires_in":3600}', 
+        headers: response_headers }
+    end
+
+    def network_authentication_oicd_response
+      { 
+        body: '{"auth_req_id":"' + network_authentication_auth_request_id + '", "expires_in":120, "interval": "2"}', 
+        headers: response_headers }
     end
 
     def response